+++ paper_admin/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "3.2.0"
+__version__ = "4.0.0"
 default_app_config = "paper_admin.apps.Config"