+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## [4.0.0](https://github.com/dldevinc/paper-admin/tree/v4.0.0) - 2022-01-18
+### ⚠ BREAKING CHANGES
+- Removed support for deprecated `NullBooleanField`.
+### Features
+- Added support for Python 3.10, Django 4.0
+- Use `sass:math`.
+### Bug Fixes  
+- Fixed `paper_admin.patches.post_office`. 
+- Fixed `autocomplete_fields` in Django 3.2
+
 ## [3.2.0](https://github.com/dldevinc/paper-admin/tree/v3.2.0) - 2021-11-26
 ### Features
 - Update to Bootstrap 4.6.1